All up to 16 auxillary GPIO signals (AUX_*_GPIO) to be defined
The 'aux' bits in the struct MotionSegment that is enqueued to the PRU
is defined as an uint16_t. Make use of the entire size of the variable
and allow up to 16 GPIO signals to be defined by the hardware. Default
all undefined AUX_*_GPIO signals to GPIO_NOT_MAPPED.

The 'aux' bits are only set right before generating the motion profile
so the additional overhead to process all 16 bits should not effect the
actual motion.

Fix the variables used for the aux bits in the various structures so
they are all 16-bit.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>