VGEN5: update aux/pwm pin mapping
The aux/pwm mapping is now handled by the configuration file. Update the
VGEN5 pin mapping header to allow all the output pins to be configured
as aux gpios or pwm outputs (if applicable). The actual pin function
depends on the pinmux setting.

Include the unrouted/unused pins in the aux pin map so I don't forget
they are available in the future.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>