Add a Spindle control class.
Currently this just parses the configuration file [ Spindle ] section then
initializes the spindle 'type'.

Two spindle types are supported: simple-pwm and pololu-smc. Specify the
type with:

  [ Spindle ]
  type = simple-pwm         # or 'pololu-smc'

The simple-pwm spindle uses the LogicOutput signals that were mapped by:

  [ Aux-Mapping ]
  aux_x = spindle           # turns on/off the spindle
  aux_x = spindle-dir       # sets the cw/ccw direction

  [ PWM-Mapping ]
  pwm_x =  spindle-speed    # sets the pwm speed

All the mappings are optional but usually spindle and/or spindle-speed
will be needed.

The pololu-smc spindle uses a virtual comm port to communicate with the
Pololu Simple Motor Controller. Specify the port with:

  [ Spindle ]
  port = /dev/ttyACM0       # this is the default virtual comm port used

The pololu-smc spindle will also use the spindle LogicOutput to enable/
disable power to the controller.

The maximum RPM of the spindle should be set in the configuration file with:

  [ Spindle ]
  max_rpm = 3000           # this is the default

This will be used to limit the speed passed in the S paramter with M3/M4.
It's also used to scale the output used to set the speed.

Automatic delays can also be specified with:

  [ Spindle ]
  pwr-delay = 1           # delay (in seconds) after enableing power to the pololu-smc
  on-delay = 1            # delay (in seconds) after turning the spindle on
  off-delay = 1           # delay (in seconds) after turning the spindle off

TODO:

The Spindle object needs to be passed to the machine_control (somehow) and
GCodeMachineControl::Impl::aux_bit_commands() needs to be modified to call
the On and Off methods to control the spindle.

Also, the Spindle class needs to be able to update the aux_bits_ as done
by GCodeMachineControl::Impl::set_output_flags().

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>