Add ramping to the simple-pwm spindle implementation.
A large delta change, or direction change, could damage the FET used
to drive the spindle due to the large current draw of the motor. Add
some simple ramping to the pwm signal to minimize the problem.

The ramping is controlled by the kRampEpsilon and kRampDelay constants
and is set to 10% pwm changes every 10 milliseconds. If necessary these
could be changed to configuration options.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>