Add automatic fan control.
Add configuration options to enable automatic fan control. The "auto_fan_pwm"
option is used to automatically enable the fan at a given PWM value when a
GCode program is first started. The "auto_fan_disable_seconds" option is then
used to disable the fan after the motors are automatically disabled when the
system is idle.

NOTE: Currently the fan is only automatically disabled as long as the stream
is still open. If the stream is closed before the system is idle longer than
the "auto_motor_disable_seconds" + "auto_fan_disable_seconds" only the motors
will be disabled.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>