Improve planner for threshold angle acceleration.
Each move will try to accel to it's desired speed, move, then decel
to the desired speed of the next move. The "knots" between all the
moves have the same speeds so there are no pulse rate glitches between
moves.

If the desired speed is too high for the move, it's final speed will
be the actual obtained speed. The next move will then start from that
speed.

The only remaining issue is if we have accelerated to a speed higher
than the upcoming move can decelerate from. We need a longer look ahead
in the planner so we can deterime if the accelerated move is going to
need to slow down.

Another possible fix might be to increase the deceleration rate by some
factor.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>