planner: make segment acceleration work for 3D moves
The planner currently calculates the 2D XY angle of each move using
atan2f and the length of the X and Y components of the move. Due to
this it will only accelerate an 2D XY move if the angle between the
move and upcoming move is within the threshold tolerance.

Rework the math to use acosf and the dot product of the vectors to
find the XYZ 3D angle between the move and the upcoming move. This
will allow the planner to accelerate 3D XYZ moves.

All the planner_test tests pass with the same output.