Improve accuracy of move_to_endstop()
This function is used to find the limit endstops when homing and
for the G30 Z axis probe.

For homing the 0.5mm move steps work fine because we are usually
trying to push in a button. The 0.1mm backoff steps then walk out
of the button to set the "home" position.

For the G30 probe the 0.5mm step is to large and can result in the
probe digging into the surface. Unlike the homing move, a G30 probe
does not do a backoff. Use that to reduce the step distance.