gcode-parser: support expressions and binary/unary operations
This makes the parser support expressions based on the NIST RS274/NGC
specification. An example of an expression is:

    [1 + acos[0] - [#3 ** [4.0/2]]]

The following binary operations are supported in expressions:

  Basic mathematical operations:
    +  addition
    -  subtraction
    *  multiplication
    /  division

  Logical operations:
    OR   non-exclusive or
    XOR  exclusive or
    AND  logical and

  Other operations:
    MOD  modulus
    **   power

These unary operations are supported:

  ABS    absolute value
  ACOS   arc cosine
  ASIN   arc sine
  ATAN   arc tangent
  COS    cosine
  EXP    e raised to the given power
  FIX    round down
  FUP    round up
  LN     natural logarithm
  ROUND  round to the neasert whole number
  SIN    sine
  SQRT   square root
  TAN    tangent

Arguments to unary operations which take angle measurements are in degrees.
Values returned by unary operations which return angle measurements are also
in degrees.

Updated tests.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>