o Add binary comparison operators
  EQ or ==   Equal
  NE or !=   Not equal
  GT or >    Greater than
  GE or >=   Greater than or equal
  LT or <    Less than
  LE or <=   Less than or equal

Example:

  #2 = [[#1 >= 0] AND [#1 LE 1000]]  (True if the value of #1 is 0 to 1000)

All operators return 1.0f for true and 0.0f for false.

o Add support for 'IF/THEN/ELSEIF/THEN/ELSE'

Used to conditionally "set parameter".

Every 'IF' or 'ELSEIF' must have a [conditional] followed by a THEN.
The final 'ELSE' is optional.

The entire 'IF' statement needs to be on a single line. And there can be
no trailing codes. Comments are allowed in the front and are parsed normally.
Any trailing comment is ignored.

Example:
(set parameter 2 to -1,0,1 based on the sign of parameter 1)
IF [#1 < 0] THEN #2=-1 ELSIF [#1==0] THEN #2=0 ELSE #2=1