o gcode-parser: add initial support for WHILE loops
This currently only supports a single level of looping. Syntax is:

WHILE [conditional expression is true] DO
(gcode to be executed in loop)
(make sure to do something to cause the expression above to be false)
END

Example:

 #1=0
 #2=1
WHILE [#1 < 10] DO
 #2=[#2+#2]
 #1=[#1+1]
END

This will run 10 loops and result in parameter 2 being 1024.

Update the polygon-24-vs-arc.gcode to use a loop to create the polygon.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>