gcode-machine-control: separate the endstop and probe moves
The move to an endstop is always unlimited until the switch is
triggered. It then has a backoff until the switch is not triggered.

Probe moves are limited to the max range of the axis and do not
have a backoff.

Separate the two moves and use the new "probe" switch-mapping for
the probe switch.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>