gcode-parser: allow parameters to be queried.
Setting a parameter is done with "#<name>=<value>".

Examples:

This allows the parameters to be queried by using a set parameter command
without the '='.
Note: this functionality is not intended to be used in a GCode program but
helps while debugging.

Examples:
// #1 = 10.000000
// #index = 5.000000
// #paramwithspaces = 4.30000

This worked in the original c code but appears to have got dropped
in the c++ conversion.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>