machine-control: don't bail if spindle cannot be initialized
Currently, if the spindle.Init() fails when machine-control starts the progam
terminates.

The 'simple-pwm' type always returns true from its Init() but the "pololu-smc"
type will return false if the USB virtual serial port cannot be opened.

Modify the code so that it doesn't bail if the spindle cannot be initialized.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>