gcode-machine-control: add helpers to turn on/off the spindle
The PRU queue needs to be flushed before turning on/off the spindle
to ensure that motion has stopped.

For clarity, add some helper functions to turn on/off the spindle
and make sure they do a planner_->BringPathToHalt() first.

Now that we have the guarantee that we statically modify the aux bits,
modify set_output_flags() to make sure that the GPIO outputs are
set. This makes the delays work correctly. It also fixes the
PololuSMCSpindle implementation when it uses a "spindle" output
to enable the power to the SMC module.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>