gcode-machine-control: wait for pause switch to clear
If the machine configuration has a pause switch, the switch will be
checked after a dwell() if it has been enabled by an M120 command.
If the switch is active wait_for_start() will be called. If the
machine also has a start switch, wait_for_start() will then block
until the switch is active.

If the pause switch is actually something like a light curtian the
wait_for_start() needs to wait until the pause switch is disabled
before waiting for the start switch.

Fix wait_for_start() to ensure that the pause switch is disabled.
Add some feedback to the stream about the pause and start switch.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>