gcode-parser: reorder the GCodeParserAxis enum
Move AXIS_E to the end of the enum. This makes the axis enumeration
a bit more logical:

X/Y/Z are the "normal" cartesian axes
A/B/C are the "rotary" axes in the X/Y/Z planes
U/V/W are the "parallel" axes in the X/Y/Z planes
E is the Extruder axis

For completeness, also reorder the switch statements associated with
this enum.

Of course, the "real" use of all the axes is determined by the user.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>