motor-operations: fix the driver_flip_ issues caused by the M114 changes
The only place the driver_flip_ needs to be applied is when the direction_bits
are being set when a new segment is being added to the PRU queue.

This fixes the issues with SetExternalPosition(), GetCurrentPosition(), and
AssignMotorSteps() that showed up with the M114 changes that used the
motor-operations GetPhysicalStatus backend.

The issues only seem to show up with negative motor directions and/or when
the home position for an axis is 'max'.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>