Use 'double' instead of 'float' for most of motion planner.
In particular with high resolution steppers (5000 steps/mm), sometimes
the planner generated unnecessary slowdowns in shallow angles, probably
due to rounding errors. In particular since some of these calculations
require squares of larger values, this resulted in rounding issues if
operated in float.

Anyway, 'double' resolution shall be good for everyone :)

Determined speed impact by running gcode-print-stats on a bunch
of files and the runtime was the same within noise.