gcode-machine-control: make the soft/hard e-stop work
For safety, an e-stop (soft or hard) should disable the motors and any
aux outputs. The e-stop event then needs to be cleared and the machine
re-homed before motion should be allowed.

For this initial implementation, the hard estop is checked after every
dwell to avoid reading the gpio while the PRU may be emmiting steps.
It is also checked when the GCodeStreamer is idle.

Currently a G28 will automatically clear the e-stop event, as long as the
hard e-stop is cleared, and re-home the machine. A M999 will also clear
the event (if the hard estop is clear) but a G28 is still needed to re-home
the machine.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>