fd-mux: fix stopping program using a systemd service
Using systemd and a startup script to start machine-control causes two
SIGTERM signals to be sent. The first one is caught by the sigaction
but the current SA_RESETHAND flag causes the second one to be treated
as a SIGKILL. This results in machine-control terminating without
properly shutting down.

Just use '0' for the sa.sa_flags so that the default signal handler
is not restored after getting the signal.

Thanks to Henner Zeller for the insight to this issue.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>