Detecting the anomaly acceleration in a segment
I found a problem with the wrong speed determining for the segment in the planner. This occurs with short segments of the movement,followed by a change in the direction of motion. This is typical for the case when a straight line goes into an arc or circle. In these segments, due to a high initial speed and a short braking path in the segment,there is an acceleration exceeding the permissible limits. That as a result leads to an uncontrollable situation.