Allow G0/G1 Z-axis moves to be clamped to the machine cube.
This adds a new configuration option, "clamp-to-range", that will allow
G0/G1 moves to be parsed that are outside the machine cube space. These
moves will be clamped to the machine cube and still produce motion.

The clamp-to-range option will allow to clamp a movement before the
range-check is testing.

While clamp-to-range takes a string of desired axes to clamp to,
for now only the Z axis is allowed if it is the only axis moving.
This makes the most sense in 2.5D applications without accidentally
causing undesirable work results.

Based on a patch by H Hartley Sweeten <hsweeten@visionengravers.com>