Add support to the PRU code to check the EStop input before each step pulse. If it's detected, update the queue status and change the queue header state to STATE_ABORT then abort the step generation for the segment. If there are more segments in the queue they will also get aborted due to the EStop input being active.
Modify the PRUMotionQueue::Enqueue() to detect the STATE_ABORT and
return false. This indicates that the motion segment could not be
added to the queue due to the abort condition caused by the EStop.

Modify all the applicable classes so that the Enqueue failure is
propogated back to the Planner so that CGodeMachineControl can
detect the abort condition and then check_for_estop(). This will
then turn all the aux outputs off, disable the motors, and set the
homing_state to NEVER_HOMED.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>