Make sure spindle is off when setting estop.
The Pololu SMC spindle controller needs to be sent a command to
ensure that the spindle is off when we detect and set the estop
condition.

The spindle _will_ turn off if the hardware is set up to disable
the spindle power when an estop is active. But, if the command is
not sent the SMC will try to automatically reenable the spindle
as soon as the estop is cleared.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>