Provide floating point parsing routines from string_view.
Unfortunately, not all c++17 compilers actually do have
std::from_chars() for floating point values available. So we're
using compile time detection and fall back to C standard library
calls if needed.

Also, rename the safe_strto*() functions to convert_strto*() and
instead of a boolean, return the const char* at the end of the
parse or nullptr in failure. This allows them to be used in
places where the input is parsed an we need to skip parsed values.